+++ BambooTracker/gui/labeled_horizontal_slider.cpp
@@ -1,5 +1,6 @@
 #include "labeled_horizontal_slider.hpp"
 #include "ui_labeled_horizontal_slider.h"
+#include <QWheelEvent>
 #include "slider_style.hpp"
 
 LabeledHorizontalSlider::LabeledHorizontalSlider(QWidget *parent) :
@@ -22,6 +23,7 @@ void LabeledHorizontalSlider::init(QString text)
 	ui->textLabel->setText(text);
 	ui->valueLabel->setText(QString::number(ui->slider->value()));
 	ui->slider->setStyle(new SliderStyle());
+	ui->slider->installEventFilter(this);
 }
 
 LabeledHorizontalSlider::~LabeledHorizontalSlider()
@@ -67,6 +69,20 @@ QString LabeledHorizontalSlider::text() const
 void LabeledHorizontalSlider::setText(QString text)
 {
 	ui->textLabel->setText(text);
+}
+
+bool LabeledHorizontalSlider::eventFilter(QObject* watched, QEvent* event)
+{
+	if (watched == ui->slider) {
+		if (event->type() == QEvent::Wheel) {
+			auto e = dynamic_cast<QWheelEvent*>(event);
+			if (e->angleDelta().y() > 0) ui->slider->setValue(ui->slider->value() + 1);
+			else if (e->angleDelta().y() < 0) ui->slider->setValue(ui->slider->value() - 1);
+			return true;
+		}
+	}
+
+	return QFrame::eventFilter(watched, event);
 }
 
 void LabeledHorizontalSlider::on_slider_valueChanged(int value)