+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix default place of load/save instrument
 - [#25] - Change FM synthesis frequency from 110933Hz to 55466Hz (thanks [@papiezak])
 - [#6], [#14] - Fix to set new property when adding instrument (thanks [@ScoopJohn], [@jimbo1qaz])
+- [#2] - Fix wheel step in slider (thanks [@pedipanol])
 
 [#25]: https://github.com/rerrahkr/BambooTracker/issues/25
 [#14]: https://github.com/rerrahkr/BambooTracker/issues/14